강의 등록 페이지
강의 등록 페이지 레포지토리/서비스/컨트롤러 작성
templates 부분 재구성

+++ sc/src/main/java/com/lms/sc/repository/NoteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lms.sc.entity.Note;
 
-public interface NoteRepository extends JpaRepository<Note, Long> {
+public interface NoteRepository extends JpaRepository<Note, Integer> {
 
 }
